aicells-python/aicells_pkg/requirements.txt
xlwings==0.17.1
scikit-learn==0.22.1
statsmodels==0.10.1
pandas==0.25.3
numpy==1.18.1
pyyaml==5.3
xgboost==0.90